services/dl-api/requirements.txt
# FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Computer Vision and Deep Learning
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.26.0
scipy>=1.11.0

# Deep Learning Frameworks
torch>=2.1.0
torchvision>=0.16.0
tensorflow>=2.15.0
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Data Processing
pandas>=2.1.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and Monitoring
loguru==0.7.2

# Image Processing
imageio==2.31.6
matplotlib==3.8.2
